pdf_parser/requirements.txt
PyPDF2==3.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0 